requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
flask-wtf
email_validator